Created Basket, Series, Inventory, BasketItem API

+++ src/main/java/com/qrzlo/webshop/data/repository/SeriesRepository.java
@@ -4,7 +4,7 @@ import com.qrzlo.webshop.data.domain.Product;
 import com.qrzlo.webshop.data.domain.Series;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SeriesRepository extends CrudRepository<Series, Long>
+public interface SeriesRepository extends CrudRepository<Series, Integer>
 {
 	Series findSeriesByProducts(Product product);
 }
